alter topic related endpoints to reflect api contract
