+++ src/CapabilityService.WebApi/Features/Kafka/Infrastructure/Api/DTOs/Topic.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Linq;
+
+namespace DFDS.CapabilityService.WebApi.Infrastructure.Api.DTOs
+{
+    public class Topic
+    {
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public Guid CapabilityId { get; set; }
+
+        public static Topic CreateFrom(Domain.Models.Topic topic)
+        {
+            return new Topic
+            {
+                Name = topic.Name,
+                Description = topic.Description,
+                CapabilityId = topic.CapabilityId,
+            };
+        }
+    }
+}
