+++ src/CapabilityService.WebApi/Features/Kafka/Infrastructure/Api/DTOs/TopicInput.cs
@@ -1,0 +1,8 @@
+﻿namespace DFDS.CapabilityService.WebApi.Infrastructure.Api.DTOs
+{
+    public class TopicInput
+    {
+        public string Name { get; set; }
+        public string Description { get; set; }
+    }
+}
